Renamed UnsupportedFooterControl to FooterMessageControl
